requirements.txt
transformers>=4.50.0
datasets>=3.5.0
torch>=2.0.0
numpy>=1.17.0
pandas>=2.0.0
peft>=0.5.0
tqdm>=4.66.0
wandb>=0.19.0
matplotlib>=3.5.0
scikit-learn>=1.0.0
jupyter>=1.0.0
nltk>=3.8.1
python-Levenshtein>=0.23.0
tensorboard>=2.14.0
protobuf>=3.20.3
huggingface-hub>=0.20.2
accelerate>=0.26.0
colorama>=0.4.6 